Refactored dkim
